django-site/django-site/requirements.txt
Django>=2.0
django-allauth>=0.36
django-countries
django-mailer>=1.2.4
mysqlclient
python-memcached
elasticsearch-dsl>=6.0,<7.0
#elasticsearch-dsl>=5.0.0,<6.0.0
python-decouple
django-storages[boto3]>=1.6.6
#boto3
#django-storages>=1.1.3
channels
Twisted[tls,http2]
djangorestframework
pinax-templates
pinax-messages
django-bootstrap4
markdown
django-filter
pygments
python-logstash